refactor: make ViewModel the authority on dialog-open guards, extract dialog action enums
  FABs into showManualEntryDialog(), confirmBookSettlement(), showAddDialog()
  so the ViewModel is the single source of truth
- Extract AccountingDialogAction -> AccountingDialogAction.kt
  and InventoryDialogAction -> InventoryDialogAction.kt so Screen
  files don't import ViewModel internals
