+++ app/src/main/java/com/haertibraeu/hopledger/ui/accounting/AccountingDialogAction.kt
@@ -1,0 +1,7 @@
+package com.haertibraeu.hopledger.ui.accounting
+
+enum class AccountingDialogAction {
+    MANUAL_ENTRY,
+    DELETE_ENTRY,
+    BOOK_SETTLEMENT,
+}
