+++ app/src/main/java/com/haertibraeu/hopledger/ui/inventory/InventoryDialogAction.kt
@@ -1,0 +1,14 @@
+package com.haertibraeu.hopledger.ui.inventory
+
+enum class InventoryDialogAction {
+    ADD_CONTAINER,
+    MOVE,
+    FILL,
+    RESERVE,
+    UNRESERVE,
+    SELL,
+    SELF_CONSUME,
+    RETURN,
+    DESTROY_BEER,
+    DELETE,
+}
